Update HF BERT test to forward-backward
