--- test/local_test_forward_hf_bert.py
@@ -1,143 +1,0 @@
-# Copyright (c) Meta Platforms, Inc. and affiliates
-import argparse
-import inspect
-import os
-import socket
-from typing import Dict
-
-import torch
-import torch.distributed.rpc as rpc
-import torch.multiprocessing as mp
-
-import transformers.utils.fx as fx
-from pippy.IR import MultiUseParameterConfig, Pipe, PipeSplitWrapper, annotate_split_points
-from pippy.PipelineDriver import PipelineDriverFillDrain, PipelineDriver1F1B, PipelineDriverBase
-from pippy.microbatch import TensorChunkSpec
-from transformers import BertModel, BertConfig
-import logging
-
-
-PROFILING_ENABLED = True
-CHECK_NUMERIC_EQUIVALENCE = True
-
-schedules = {
-    'FillDrain': PipelineDriverFillDrain,
-    '1F1B': PipelineDriver1F1B,
-}
-
-VERBOSE = bool(int(os.environ.get('VERBOSE', False)))
-
-if VERBOSE:
-    logging.getLogger().setLevel(logging.DEBUG)
-
-@torch.fx.wrap
-def torch_ones_wrapper(*args, **kwargs):
-    return torch.ones(*args, **kwargs)
-
-
-class HFBertTracer(fx.HFTracer):
-    def trace(self, root, concrete_args=None, method_names=None):
-        graph = super().trace(root, concrete_args, method_names)
-        # HACK to replace HF's non-resolvable wrappers with the original
-        # tensor constructor functions
-        # Requires this patch to HF: https://github.com/jamesr66a/transformers/commit/ede9d30f36f12be390692617ef76e2928b1612bd
-        for node in graph.nodes:
-            if node.op == 'call_function':
-                if getattr(node.target, '_orig', None) == torch.ones:
-                    node.target = torch_ones_wrapper
-        return graph
-
-
-torch.fx.Tracer.proxy_buffer_attributes = True
-
-
-def run_master(args):
-    bs = 20
-    seq_length = 32
-
-    MULTI_USE_PARAM_CONFIG = MultiUseParameterConfig.REPLICATE if args.replicate else MultiUseParameterConfig.TRANSMIT
-    print(f'REPLICATE config: {args.replicate} -> {MULTI_USE_PARAM_CONFIG}')
-
-    print("Using schedule:", args.schedule)
-
-    bert = BertModel(BertConfig())
-    bert.eval()
-    bert_input = torch.zeros(bs, seq_length, dtype=torch.long).random_(bert.config.vocab_size)
-    bert(bert_input)
-
-    for i in range(bert.config.num_hidden_layers):
-        annotate_split_points(bert, {f'encoder.layer.{i}': PipeSplitWrapper.SplitPoint.BEGINNING})
-    annotate_split_points(bert, {'pooler': PipeSplitWrapper.SplitPoint.BEGINNING})
-
-    input_names = bert.dummy_inputs.keys()
-    sig = inspect.signature(bert.forward)
-    concrete_args = {p.name: p.default for p in sig.parameters.values() if p.name not in input_names}
-
-    hf_tracer = HFBertTracer()
-
-    print('Instantiating BERT Pipeline')
-    bert_pipe = Pipe.from_tracing(bert, MULTI_USE_PARAM_CONFIG, tracer=hf_tracer, concrete_args=concrete_args)
-
-    assert bert.config.num_hidden_layers + 2 == len(list(bert_pipe.split_gm.children()))
-
-    args_chunk_spec = (TensorChunkSpec(0),)
-    kwargs_chunk_spec: Dict = {}
-    output_chunk_spec = {'last_hidden_state': TensorChunkSpec(0), 'pooler_output': TensorChunkSpec(0)}
-
-    pipe_driver: PipelineDriverBase = schedules[args.schedule](bert_pipe, args_chunk_spec, kwargs_chunk_spec,
-                                                               output_chunk_spec,
-                                                               args.world_size, _debug_mask_minibatches=True)
-
-    # # Warm up and correctness runs
-    out = pipe_driver.run(5, bert_input)
-    ref_out = bert_pipe(bert_input)
-
-    if CHECK_NUMERIC_EQUIVALENCE:
-        torch.testing.assert_close(out['last_hidden_state'], ref_out['last_hidden_state'])
-        torch.testing.assert_close(out['pooler_output'], ref_out['pooler_output'])
-        print(
-            f'equivalence test passed {torch.sum(out["last_hidden_state"])} ref {torch.sum(ref_out["last_hidden_state"])}')
-
-    # # Profiling runs
-    with torch.autograd.profiler_legacy.profile(enabled=PROFILING_ENABLED) as prof:
-        pipe_driver._debug_mask_minibatches = False
-        out = pipe_driver.run(5, bert_input)
-        ref_out = bert_pipe(bert_input)
-        print(
-            f'profiling run completed {torch.sum(out["last_hidden_state"])} ref {torch.sum(ref_out["last_hidden_state"])}')
-    if PROFILING_ENABLED:
-        prof.export_chrome_trace(f'{os.path.splitext(os.path.basename(__file__))[0]}.json')
-
-
-def run_worker(rank, world_size, args):
-    print(f"rank = {rank} host/pid = {socket.gethostname()}/{os.getpid()}")
-    os.environ['MASTER_ADDR'] = args.master_addr
-    os.environ['MASTER_PORT'] = args.master_port
-    options = rpc.TensorPipeRpcBackendOptions(num_worker_threads=256)
-    rpc.init_rpc(
-        f"worker{rank}",
-        rank=rank,
-        world_size=world_size,
-        rpc_backend_options=options
-    )
-    if rank == 0:
-        run_master(args)
-    rpc.shutdown()
-
-
-if __name__ == "__main__":
-    parser = argparse.ArgumentParser()
-    parser.add_argument('--world_size', type=int, default=int(os.getenv("WORLD_SIZE", 14)))
-    parser.add_argument('--rank', type=int, default=int(os.getenv("RANK", -1)))
-    parser.add_argument('--master_addr', type=str, default=os.getenv('MASTER_ADDR', 'localhost'))
-    parser.add_argument('--master_port', type=str, default=os.getenv('MASTER_PORT', '29500'))
-    parser.add_argument('-s', '--schedule', type=str, default=list(schedules.keys())[0], choices=schedules.keys())
-    parser.add_argument('--replicate', type=int, default=int(os.getenv("REPLICATE", '0')))
-    args = parser.parse_args()
-
-    if args.rank == -1:
-        mp.spawn(run_worker, args=(args.world_size, args,), nprocs=args.world_size, join=True)
-    elif args.rank < args.world_size:
-        run_worker(args.rank, args.world_size, args)
-    else:
-        print("I'm unused, exiting")
